refactor: update import paths for Taxonomy and Term classes to reflect new directory structure
